Chop TCP in IPsec output.

+++ netinet/tcp_var.h
@@ -715,8 +715,7 @@ struct tcpcb *
 	 tcp_newtcpcb(struct inpcb *, int);
 void	 tcp_notify(struct inpcb *, int);
 int	 tcp_output(struct tcpcb *);
-int	 tcp_chopper(struct mbuf *, struct mbuf_list *, struct ifnet *,
-	     u_long);
+int	 tcp_chopper(struct mbuf *, struct mbuf_list *, struct ifnet *, u_int);
 void	 tcp_pulloutofband(struct socket *, u_int, struct mbuf *, int);
 int	 tcp_reass(struct tcpcb *, struct tcphdr *, struct mbuf *, int *);
 void	 tcp_rscale(struct tcpcb *, u_long);
